LPD-69514 Reorganized SSA Dashboard folder
