+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/SSADashboard/SSADashboardRouter.tsx
@@ -7,9 +7,9 @@ import {HashRouter, Route, Routes} from 'react-router-dom';
 
 import withProviders from '../../hoc/withProviders';
 import SSADashboardOutlet from './SSADashboardOutlet';
-import MySaaSTrials from './pages';
+import MySaaSTrials from './pages/MySaaSTrials';
 import SaaSTrials from './pages/SaaSTrial';
-import TrialDetails from './pages/trialDetails';
+import TrialDetails from './pages/TrialDetails';
 
 import './index.scss';
 
@@ -18,7 +18,9 @@ const SSADashboardRouter = () => (
 		<Routes>
 			<Route element={<SSADashboardOutlet />}>
 				<Route element={<MySaaSTrials />} index />
+
 				<Route element={<SaaSTrials />} path="saas-trials" />
+
 				<Route element={<TrialDetails />} path="details/:orderId" />
 			</Route>
 		</Routes>
