+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/SSADashboard/pages/TrialDetails/TrialDetailsBody.tsx
